re-use strip_colors
